+++ includes/i18n/languages.php
@@ -1,6 +1,6 @@
 <?php
-    // File Name => Language Name
 
+    // File Name => Language Name
     $languages = [
         // English first
         "en" => "English",
@@ -9,12 +9,13 @@
         "el" => "Ελληνικά",
         "es" => "Español",
         "fr" => "Français",
+        "it" => "Italiano",
         "jp" => "日本語",
-	"pl" => "Polski",
+        "pl" => "Polski",
         "pt" => "Português",
         "pt_br" => "Português Brasileiro",
-	"sr" => "Српски",
-	"sr_lat" => "Srpski",
+        "sr_lat" => "Srpski",
+        "sr" => "Српски",
         "tr" => "Türkçe",
         "zh_cn" => "简体中文",
         "zh_tw" => "繁體中文",