+++ includes/inputvalidation.php
@@ -4,7 +4,6 @@ function validate($value) {
     $value = trim($value);
     $value = stripslashes($value);
     $value = htmlspecialchars($value);
-    $value = htmlentities($value);
     return $value;
 }